… IMPLEMENTATION_VERIFICATION_REPORT.md
… apps/api/src/server.ts
… specs/006-you-are-a/tasks.md
… tests/integration/migration-status.test.ts