### **User description**
…onfiguration

- Increased test timeout in vitest configuration for performance tests.
- Added comprehensive implementation verification report for specification 006-you-are-a, confirming successful completion of all tasks and compliance with Brazilian healthcare regulations.
- Implemented integration tests for architecture configuration API, validating end-to-end functionality, database integration, and healthcare compliance.
- Created migration status API integration tests to track migration progress and validate healthcare compliance.


___

### **PR Type**
Enhancement, Tests, Documentation


___

### **Description**
• **Added comprehensive API endpoints** for architecture configuration, migration status tracking, performance metrics, and healthcare compliance validation
• **Implemented Bun-optimized database client** with Supabase integration, healthcare compliance features (LGPD/ANVISA/CFM), and performance monitoring
• **Created extensive integration test suite** covering architecture configuration API, migration status tracking, and end-to-end functionality validation
• **Enhanced test configuration** by increasing timeout for performance tests from 60s to 120s
• **Added comprehensive implementation verification report** documenting 100% completion of 191 tasks with Brazilian healthcare compliance validation
• **Updated task completion status** marking all architecture analysis phases as completed


___

### Diagram Walkthrough


```mermaid
flowchart LR
  A["API Server"] --> B["Architecture Config API"]
  A --> C["Migration Status API"]
  A --> D["Performance Metrics API"]
  A --> E["Compliance Status API"]
  F["Database Client"] --> G["Supabase Integration"]
  F --> H["Healthcare Compliance"]
  F --> I["Bun Runtime Optimization"]
  J["Integration Tests"] --> B
  J --> C
  K["Implementation Report"] --> L["100% Task Completion"]
  K --> M["Healthcare Compliance Validation"]
```



<details> <summary><h3> File Walkthrough</h3></summary>

<table><thead><tr><th></th><th align="left">Relevant files</th></tr></thead><tbody><tr><td><strong>Enhancement</strong></td><td><table>
<tr>
  <td>
    <details>
      <summary><strong>server.ts</strong><dd><code>Add comprehensive API endpoints for architecture configuration and </code><br><code>compliance</code></dd></summary>
<hr>

apps/api/src/server.ts

• Added comprehensive Architecture Configuration API endpoints <br>(<code>/api/architecture/config</code>) with Bun runtime, healthcare compliance, <br>and performance optimization settings<br> • Implemented Migration Status <br>API endpoints (<code>/api/migration/status</code>) to track migration progress <br>through phases with healthcare compliance validation<br> • Added <br>Performance Metrics API endpoints (<code>/api/performance/*</code>) for build, <br>runtime, edge, memory, and healthcare benchmark monitoring<br> • Created <br>Compliance Status API endpoints (<code>/api/compliance/*</code>) for LGPD, ANVISA, <br>CFM, data residency, and security audit validation


</details>


  </td>
  <td><a href="https://github.com/GrupoUS/neonpro/pull/79/files#diff-3504cff00dc0b301e00da4235ed65736b5fd11064dd71e70f2cca2e4f00f77eb">+937/-0</a>&nbsp; </td>

</tr>

<tr>
  <td>
    <details>
      <summary><strong>client.ts</strong><dd><code>Implement comprehensive Bun-optimized database client with healthcare </code><br><code>compliance</code></dd></summary>
<hr>

packages/database/src/client.ts

• Replaced minimal stub implementation with comprehensive Supabase <br>database client optimized for Bun runtime<br> • Added healthcare <br>compliance features (LGPD/ANVISA/CFM) with configuration schema <br>validation using Zod<br> • Implemented performance monitoring, connection <br>pooling, and edge runtime support with healthcare-grade security<br> • <br>Added migration tracking, health checking, and Bun-specific <br>optimizations for database operations


</details>


  </td>
  <td><a href="https://github.com/GrupoUS/neonpro/pull/79/files#diff-f10fbd6f295860a8f0d402f2d9fd9ea8dafa20e10ea2efed6c3017b32135593e">+427/-8</a>&nbsp; </td>

</tr>
</table></td></tr><tr><td><strong>Tests</strong></td><td><table>
<tr>
  <td>
    <details>
      <summary><strong>architecture-config.test.ts</strong><dd><code>Add integration tests for architecture configuration API endpoints</code></dd></summary>
<hr>

tests/integration/architecture-config.test.ts

• Created comprehensive integration tests for Architecture <br>Configuration API endpoints<br> • Added validation tests for Bun runtime <br>configuration, performance optimization, and edge runtime settings<br> • <br>Implemented healthcare compliance validation tests for LGPD, ANVISA, <br>and CFM requirements<br> • Added database integration tests to validate <br>end-to-end functionality and configuration persistence


</details>


  </td>
  <td><a href="https://github.com/GrupoUS/neonpro/pull/79/files#diff-3253b1fe986127b734f82d1aefc727a6637524e45c26a2bd8291681bc3904340">+262/-0</a>&nbsp; </td>

</tr>

<tr>
  <td>
    <details>
      <summary><strong>migration-status.test.ts</strong><dd><code>Add integration tests for migration status API endpoints</code>&nbsp; </dd></summary>
<hr>

tests/integration/migration-status.test.ts

• Created integration tests for Migration Status API endpoints with <br>phase tracking validation<br> • Added migration progress tracking tests <br>and configuration validation with performance metrics<br> • Implemented <br>healthcare compliance testing for migration processes and database <br>integration validation<br> • Added comprehensive metrics testing for <br>baseline, current, and target performance measurements


</details>


  </td>
  <td><a href="https://github.com/GrupoUS/neonpro/pull/79/files#diff-023359b255ab8e7b30926f732557e18af8fe79e73fd1a0157d455dedbfce900d">+237/-0</a>&nbsp; </td>

</tr>
</table></td></tr><tr><td><strong>Configuration changes</strong></td><td><table>
<tr>
  <td>
    <details>
      <summary><strong>vitest.config.ts</strong><dd><code>Update vitest configuration for performance test timeout</code>&nbsp; </dd></summary>
<hr>

vitest.config.ts

• Increased test timeout from 60000ms to 120000ms for performance <br>tests<br> • Removed duplicate <code>maxWorkers</code> and <code>minWorkers</code> configuration <br>entries


</details>


  </td>
  <td><a href="https://github.com/GrupoUS/neonpro/pull/79/files#diff-2ee894bf23aa44ff4ce12a3da8af19ab20180474ebf2176dbe5f2eea3f96dc92">+1/-3</a>&nbsp; &nbsp; &nbsp; </td>

</tr>
</table></td></tr><tr><td><strong>Documentation</strong></td><td><table>
<tr>
  <td>
    <details>
      <summary><strong>IMPLEMENTATION_VERIFICATION_REPORT.md</strong><dd><code>Add comprehensive implementation verification report with production </code><br><code>readiness assessment</code></dd></summary>
<hr>

IMPLEMENTATION_VERIFICATION_REPORT.md

• Added comprehensive implementation verification report documenting <br>successful completion of all 54 tasks (T001-T054)<br> • Documented 100% <br>completion rate across all phases with detailed technical verification <br>results<br> • Included Brazilian healthcare compliance validation <br>(LGPD/ANVISA/CFM) and performance metrics validation<br> • Provided <br>production readiness assessment with approval for immediate deployment <br>to Brazilian aesthetic clinics


</details>


  </td>
  <td><a href="https://github.com/GrupoUS/neonpro/pull/79/files#diff-179e2583a8876acb5bf6ee745b8b61a9707b66c7e6030b844b05e7bc21d38efc">+216/-0</a>&nbsp; </td>

</tr>

<tr>
  <td>
    <details>
      <summary><strong>tasks.md</strong><dd><code>Mark comprehensive architecture analysis tasks as completed</code></dd></summary>
<hr>

specs/006-you-are-a/tasks.md

• Updated task completion status from unchecked to checked for all <br>phases (3.1-3.6)<br> • Marked 191 tasks as completed across setup, <br>testing, implementation, API, workflow, and integration phases<br> • Added <br>"✅ COMPLETED" status indicators to phase headers<br> • Maintained detailed <br>task structure and healthcare compliance requirements


</details>


  </td>
  <td><a href="https://github.com/GrupoUS/neonpro/pull/79/files#diff-83bac89bc29db23ae13f55eac6eca307629afde64b2997bfd7a09b598a35c620">+352/-321</a></td>

</tr>
</table></td></tr></tr></tbody></table>

</details>

___



<!-- This is an auto-generated comment: release notes by coderabbit.ai -->

## Summary by CodeRabbit

* New Features
  * Introduced comprehensive public APIs for architecture configuration, performance metrics, migration status/config/metrics, and compliance (including audit submission).
  * Enhanced API info to surface grouped performance and compliance capabilities.
  * Upgraded database layer with health checks, performance monitoring, and edge/runtime optimizations.

* Tests
  * Added integration tests covering architecture configuration and migration lifecycle endpoints.

* Documentation
  * Added a production-ready implementation verification report and finalized Phase 3 execution plan.

* Chores
  * Increased test timeout to improve stability and removed redundant configuration.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->